Moving test around and added tests for string equivalence classes
