+++ NTestDataBuilder/EquivalenceClasses/StringEquivalenceClasses.cs
@@ -27,5 +27,16 @@ namespace NTestDataBuilder.EquivalenceClasses
         {
             return fixture.Fixture.Create(prefix);
         }
+
+        /// <summary>
+        /// Generate and return a string ending with the given suffix.
+        /// </summary>
+        /// <param name="fixture">The fixture to generate a string for</param>
+        /// <param name="suffix">String to end the returned anonymous string with</param>
+        /// <returns>The generated string</returns>
+        public static string StringEndingWith(this AnonymousValueFixture fixture, string suffix)
+        {
+            return string.Format("{0}{1}", fixture.Fixture.Create<string>(), suffix);
+        }
     }
 }
